Fetching certified copy details from api
