+++ src/controllers/certificates/check.details.controller.ts
@@ -1,13 +1,14 @@
-import { getAccessToken, getUserId } from "../../session/helper";
 import { NextFunction, Request, Response } from "express";
-import { addItemToBasket, getCertificateItem, getBasket } from "../../client/api.client";
-import { CHS_URL, APPLICATION_NAME } from "../../config/config";
-import { CertificateItem, ItemOptions } from "ch-sdk-node/dist/services/order/item/certificate/types";
-import { CERTIFICATE_CHECK_DETAILS } from "../../model/template.paths";
-
-import { CERTIFICATE_OPTIONS, CERTIFICATE_DELIVERY_DETAILS, replaceCertificateId } from "../../model/page.urls";
+import { CertificateItem, ItemOptions } from "ch-sdk-node/dist/services/certificates/types";
 import { Basket, DeliveryDetails } from "ch-sdk-node/dist/services/order/basket/types";
 import { createLogger } from "ch-structured-logging";
+
+import { CERTIFICATE_OPTIONS, CERTIFICATE_DELIVERY_DETAILS, replaceCertificateId } from "../../model/page.urls";
+import { mapDeliveryDetails, mapToHtml, mapDeliveryMethod } from "../../utils/check-details-utils";
+import { CERTIFICATE_CHECK_DETAILS } from "../../model/template.paths";
+import { addItemToBasket, getCertificateItem, getBasket } from "../../client/api.client";
+import { CHS_URL, APPLICATION_NAME } from "../../config/config";
+import { getAccessToken, getUserId } from "../../session/helper";
 
 const GOOD_STANDING = "Statement of good standing";
 const REGISTERED_OFFICE_ADDRESS = "Registered office address";
@@ -84,54 +85,6 @@ export const mapIncludedOnCertificate = (itemOptions: ItemOptions): string => {
     }
 
     return mapToHtml(mappings);
-};
-
-export const mapDeliveryDetails = (deliveryDetails: DeliveryDetails | undefined): string => {
-    const mappings:string[] = [];
-
-    if (deliveryDetails === undefined) {
-        return "";
-    }
-
-    mappings.push(deliveryDetails.forename + " " + deliveryDetails.surname);
-    mappings.push(deliveryDetails.addressLine1);
-
-    if (deliveryDetails.addressLine2 !== "" && deliveryDetails.addressLine2 !== undefined) {
-        mappings.push(deliveryDetails.addressLine2);
-    }
-
-    mappings.push(deliveryDetails.locality);
-
-    if (deliveryDetails.region !== "" && deliveryDetails.region !== undefined) {
-        mappings.push(deliveryDetails.region);
-    }
-
-    if (deliveryDetails.postalCode !== "" && deliveryDetails.postalCode !== undefined) {
-        mappings.push(deliveryDetails.postalCode);
-    }
-
-    mappings.push(deliveryDetails.country);
-
-    return mapToHtml(mappings);
-};
-
-export const mapDeliveryMethod = (itemOptions: Record<string, any>): string | null => {
-    if (itemOptions?.deliveryTimescale === "standard") {
-        return "Standard delivery (aim to dispatch within 4 working days)";
-    }
-    if (itemOptions?.deliveryTimescale === "same-day") {
-        return "Same Day";
-    }
-    return null;
-};
-
-export const mapToHtml = (mappings: string[]): string => {
-    let htmlString: string = "";
-
-    mappings.forEach((element) => {
-        htmlString += element + "<br>";
-    });
-    return htmlString;
 };
 
 export const mapCertificateType = (certificateType: string): string => {
