+++ test/controller/certificates/options.controller.integration.test.ts
@@ -2,7 +2,7 @@ import chai from "chai";
 import sinon from "sinon";
 import ioredis from "ioredis";
 import cheerio from "cheerio";
-import { CertificateItem } from "ch-sdk-node/dist/services/order/item/certificate/types";
+import { CertificateItem } from "ch-sdk-node/dist/services//certificates/types";
 
 import * as apiClient from "../../../src/client/api.client";
 import { CERTIFICATE_OPTIONS, replaceCertificateId } from "../../../src/model/page.urls";
